DtInput renamed to DateInput, dtinput tests renamed to dateinput, DateInput legacy implementation removed.
